Add tests for new API; update src of new API
